+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Fix markdown handling of autolinks with angle brackets and factorize (and test) markdown `parse_html()` [#1625](https://github.com/opendatateam/udata/pull/1625)
 - Fix timeline order [#1642](https://github.com/opendatateam/udata/pull/1642)
 - Fix markdown rendering on IE11 [#1645](https://github.com/opendatateam/udata/pull/1645)
+- Consider bad UUID as 404 in routing [#1646](https://github.com/opendatateam/udata/pull/1646)
 
 ## 1.3.8 (2018-04-25)